Split resources in more distinct namespaces
