+++ src/backend/main/py.main/caliopen_main/common/objects/tag.py
@@ -1,0 +1,30 @@
+# -*- coding: utf-8 -*-
+"""Caliopen User tag parameters classes."""
+from __future__ import absolute_import, print_function, unicode_literals
+import types
+import uuid
+import datetime
+
+from .base import ObjectJsonDictifiable
+from ..store import ResourceTag as ModelResourceTag
+from ..store import IndexedResourceTag
+
+import logging
+log = logging.getLogger(__name__)
+
+
+class ResourceTag(ObjectJsonDictifiable):
+    """Tag nested in resources."""
+
+    _attrs = {
+        'date_insert': datetime.datetime,
+        'importance_level': types.IntType,
+        'label': types.StringType,
+        'name': types.StringType,
+        'tag_id': uuid.UUID,
+        'type': types.StringType,
+    }
+
+    _model_class = ModelResourceTag
+    _pkey_name = 'tag_id'
+    _index_class = IndexedResourceTag
