+++ src/backend/main/py.main/caliopen_main/common/parameters/tag.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+"""Caliopen tags parameters."""
+
+from schematics.models import Model
+from schematics.types import StringType, UUIDType
+
+
+class ResourceTag(Model):
+    """Tag related to a resource."""
+
+    tag_id = UUIDType()
+    name = StringType()
+    type = StringType()
