+++ src/backend/main/py.main/caliopen_main/message/objects/participant.py
