--- src/backend/main/py.main/caliopen_main/parsers/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import absolute_import, print_function, unicode_literals
-
-from .mail import MailMessage
-from .vcard import parse_vcard, parse_vcards
-
-__all__ = ['MailMessage', 'parse_vcard', 'parse_vcards']
