+++ src/backend/main/py.main/caliopen_main/user/objects/tag.py
@@ -7,15 +7,12 @@ import datetime
 
 from caliopen_main.objects import base
 from caliopen_main.user.store import UserTag as ModelUSerTag
-from caliopen_main.user.store import ResourceTag as ModelResourceTag
-from caliopen_main.user.store import IndexedResourceTag
 
 import logging
 log = logging.getLogger(__name__)
 
 
 class UserTag(base.ObjectUser):
-
     """Tag related to an user."""
 
     _attrs = {
@@ -33,21 +30,3 @@ class UserTag(base.ObjectUser):
         """Delete a tag in store."""
         self._db.delete()
         return True
-
-
-class ResourceTag(base.ObjectJsonDictifiable):
-
-    """Tag nested in resources."""
-
-    _attrs = {
-        'date_insert': datetime.datetime,
-        'importance_level': types.IntType,
-        'label': types.StringType,
-        'name': types.StringType,
-        'tag_id': uuid.UUID,
-        'type': types.StringType,
-    }
-
-    _model_class = ModelResourceTag
-    _pkey_name = 'tag_id'
-    _index_class = IndexedResourceTag
