+++ src/backend/main/py.main/caliopen_main/user/store/tag.py
@@ -5,11 +5,10 @@ from __future__ import absolute_import, print_function, unicode_literals
 import uuid
 from cassandra.cqlengine import columns
 
-from caliopen_storage.store import BaseUserType, BaseModel
+from caliopen_storage.store import BaseModel
 
 
 class UserTag(BaseModel):
-
     """User tags model."""
 
     date_insert = columns.DateTime()
@@ -19,15 +18,3 @@ class UserTag(BaseModel):
     tag_id = columns.UUID(primary_key=True, default=uuid.uuid4)
     type = columns.Text()
     user_id = columns.UUID(primary_key=True)
-
-
-class ResourceTag(BaseUserType):
-
-    """Tag nested in resource model."""
-    _pkey = 'tag_id'
-
-    importance_level = columns.Integer()
-    label = columns.Text()
-    name = columns.Text()
-    tag_id = columns.UUID()
-    type = columns.Text()
